+++ src/components/calendar/CalendarNotification.tsx
@@ -1,0 +1,30 @@
+import { Alert, AlertIcon, AlertTitle, Box, CloseButton, VStack } from '@chakra-ui/react';
+
+import { useNotifications } from '../../hooks/useNotifications';
+import { Event } from '../../types';
+
+type CalendarNotificationProps = {
+  events: Event[];
+};
+
+export const CalendarNotification = ({ events }: CalendarNotificationProps) => {
+  const { notifications, setNotifications } = useNotifications(events);
+
+  if (notifications.length === 0) return null;
+
+  return (
+    <VStack position="fixed" top={4} right={4} spacing={2} align="flex-end">
+      {notifications.map((notification, index) => (
+        <Alert key={index} status="info" variant="solid" width="auto">
+          <AlertIcon />
+          <Box flex="1">
+            <AlertTitle fontSize="sm">{notification.message}</AlertTitle>
+          </Box>
+          <CloseButton
+            onClick={() => setNotifications((prev) => prev.filter((_, i) => i !== index))}
+          />
+        </Alert>
+      ))}
+    </VStack>
+  );
+};
